style: remove custom css to avoid conflicting height
I added custom height at some point to have my custom things but
eventually I realized that it was a bad idea.
A lot of components were breaking and it was more of an hassle to fix
their styles than to remove the custom css and try to do it correctly.

Mainly, it was breaking the Select component. Opening Select was showing
a really long scroll on the brower.

Let's go for a simpler design that works and I will see afterwards if I
really need a custom css.